BREAKING CHANGE: Refactor and relocate DocxStamperConfiguration
The file DocxStamperConfiguration has been moved from org.wickedsource.docxstamper to pro.verron.officestamper.core package. Additionally, the module-info.java file has been updated to reflect these changes. Deprecated annotation is also removed from this class.
